+++ src/main/java/tp/tp_disenio_2025_grupo_28/model/Huesped.java
@@ -9,7 +9,7 @@ public class Huesped {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Integer id;
 
     private String nombre;
     private String apellido;
@@ -18,13 +18,13 @@ public class Huesped {
     private String cuit;
     private String posicionFrenteAlIva;
     private Date fechaNacimiento;
-    private int telefono;
+    private Integer telefono;
     private String email;
     private String ocupacion;
 
     // Getters y setters
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     public String getNombre() { return nombre; }
     public void setNombre(String nombre) { this.nombre = nombre; }
@@ -47,8 +47,8 @@ public class Huesped {
     public Date getFechaNacimiento() { return fechaNacimiento; }
     public void setFechaNacimiento(Date fechaNacimiento) { this.fechaNacimiento = fechaNacimiento; }
 
-    public int getTelefono() { return telefono; }
-    public void setTelefono(int telefono) { this.telefono = telefono; }
+    public Integer getTelefono() { return telefono; }
+    public void setTelefono(Integer telefono) { this.telefono = telefono; }
 
     public String getEmail() { return email; }
     public void setEmail(String email) { this.email = email; }
