+++ src/main/java/tp/tp_disenio_2025_grupo_28/repository/HuespedRepository.java
@@ -3,5 +3,5 @@ package tp.tp_disenio_2025_grupo_28.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tp.tp_disenio_2025_grupo_28.model.Huesped;
 
-public interface HuespedRepository extends JpaRepository<Huesped, Long> {
+public interface HuespedRepository extends JpaRepository<Huesped, Integer> {
 }
