+++ src/main/java/tp/tp_disenio_2025_grupo_28/repository/PersonaFisicaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import tp.tp_disenio_2025_grupo_28.model.PersonaFisica;
 
 
-public interface PersonaFisicaRepository extends JpaRepository<PersonaFisica, Long> {
+public interface PersonaFisicaRepository extends JpaRepository<PersonaFisica, Integer> {
 }
